+++ hdate/__init__.py
@@ -5,7 +5,7 @@ HDate calculates and generates a representation either in English or Hebrew
 of the Jewish calendrical date and times for a given location
 """
 
-from hdate.date import HDateInfo
+from hdate.date_info import HDateInfo
 from hdate.hebrew_date import HebrewDate, Months
 from hdate.holidays import HolidayTypes
 from hdate.location import Location
